requirements.txt
fastapi
uvicorn
pytest
httpx
python-multipart
python-docx
pdfplumber
